+++ client/src/pages/programForm/ProgramForm.js
@@ -65,7 +65,8 @@ const ProgramForm = props => {
         value => !props.programs.find(p => !!value && p.name.toLowerCase() === value.toLowerCase())
       ),
     description: yup.string()
-      .required("Description is required."),
+      .required("Description is required.")
+      .notOneOf(["<p></p>", "<p><br></p>"], "Description is required"),
     code: yup.string()
       .required("Code is required.")
       .matches("[A-Za-z0-9]+", "Code must be alphanumeric."),
@@ -83,11 +84,11 @@ const ProgramForm = props => {
       url: yup.string().url()
     }),
     attributes: yup.object()
-    .test(
-        "not empty",
-        "Attribute names must not be empty",
-        value => !Object.keys(value).find(d => d.trim() === '')
-    )
+      .test(
+          "not empty",
+          "Attribute names must not be empty",
+          value => Object.keys(value).every(d => d && d.trim() !== '')
+      ),
   });
 
   const defaultProgramValues = {
@@ -445,6 +446,7 @@ const ProgramForm = props => {
                         <AttributeInputs
                             attributes={values.attributes}
                             handleUpdate={(attributes) => setFieldValue("attributes", attributes)}
+                            error={errors.attributes}
                         />
 
                         <Row>