+++ CHANGELOG.md
@@ -14,7 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added File Manager tool to the Study Tracker UI. This tool allows the browsing of storage locations registered within Study Tracker. Users can navigate & create folders, and upload & download files.
 - Added support for Amazon Web Services S3 storage locations. This allows users to store files in S3 buckets.
 
-## [0.7.0] - 2022-XX-XX
+### Changed
+
+### Fixed
+- Fixed validation of Quill input fields to prevent empty inputs.
+- Fixed validation of required custom assay type fields.
+- Fixed validation of key-value attribute fields.
+- Fixed a number of small UI bugs.
+
+## [0.7.0] - 2022-09-26
 
 Study Tracker version 0.7.0 is the first of several transitional updates that will modernize the front and back ends of the application. These updates are primarily designed to utilize newer versions of dependencies, remove deprecated dependencies, and mitigate security vulnerabilities. A significant number of new features will be added, as well, in order to take the application in a direction that will broaden its utility.